CB-22446 Backport DL/DH os selection to 2.73
